refactor: dom-rule 상수 다 분리해라
