+++ src/app/asset/constants/BUTTON.js
@@ -1,0 +1,8 @@
+const BUTTON = {
+    addProduct: {
+        title: '추가하기',
+        id: 'product-add-button',
+    },
+};
+
+export default BUTTON;
