+++ src/app/asset/constants/INPUT_ID.js
@@ -1,0 +1,7 @@
+const INPUT_ID = {
+    productName: 'product-name-input',
+    productPrice: 'product-price-input',
+    productQuantity: 'product-quantity-input',
+};
+
+export default INPUT_ID;
