+++ src/app/asset/constants/SUB_TITLE_TEXT.js
@@ -1,0 +1,6 @@
+const SUB_TITLE_TEXT = {
+    addProduct: '상품 추가하기',
+    productInventory: '상품 현황',
+};
+
+export default SUB_TITLE_TEXT;
