+++ src/app/asset/constants/TABLE_CELL_CLASS_NAME.js
@@ -1,0 +1,9 @@
+const TABLE_CELL_CLASS_NAME = {
+    inventory: {
+        productName: 'product-manage-name',
+        productPrice: 'product-manage-price',
+        productQuantity: 'product-manage-quantity',
+    },
+};
+
+export default TABLE_CELL_CLASS_NAME;
