+++ src/app/asset/constants/TABLE_ROW_CLASS_NAME.js
@@ -1,0 +1,5 @@
+const TABLE_ROW_CLASS_NAME = {
+    inventory: 'product-manage-item',
+};
+
+export default TABLE_ROW_CLASS_NAME;
