+++ src/app/asset/constants/TABLE_TITLE.js
@@ -1,0 +1,7 @@
+const TABLE_TITLE = {
+    productName: '상품명',
+    productPrice: '가격',
+    productQuantity: '수량',
+};
+
+export default TABLE_TITLE;
