+++ src/app/asset/constants/TAP.js
@@ -1,0 +1,16 @@
+const TAP = {
+    inventory: {
+        title: '상품 관리',
+        id: 'product-add-menu',
+    },
+    coin: {
+        title: '잔돈 충전',
+        id: 'vending-machine-manage-menu',
+    },
+    purchase: {
+        title: '상품 구매',
+        id: 'product-purchase-menu',
+    },
+};
+
+export default TAP;
