+++ src/app/asset/constants/VENDING_MACHINE_TITLE.js
@@ -1,0 +1,3 @@
+const VENDING_MACHINE_TITLE = '자판기';
+
+export default VENDING_MACHINE_TITLE;
