--- src/app/asset/constants/dom-rule.js
@@ -1,52 +1,0 @@
-export const VENDING_MACHINE_TITLE = '자판기';
-
-export const TAP = {
-    inventory: {
-        title: '상품 관리',
-        id: 'product-add-menu',
-    },
-    coin: {
-        title: '잔돈 충전',
-        id: 'vending-machine-manage-menu',
-    },
-    purchase: {
-        title: '상품 구매',
-        id: 'product-purchase-menu',
-    },
-};
-
-export const INPUT_ID = {
-    productName: 'product-name-input',
-    productPrice: 'product-price-input',
-    productQuantity: 'product-quantity-input',
-};
-
-export const SUB_TITLE_TEXT = {
-    addProduct: '상품 추가하기',
-    productInventory: '상품 현황',
-};
-
-export const BUTTON = {
-    addProduct: {
-        title: '추가하기',
-        id: 'product-add-button',
-    },
-};
-
-export const TABLE_ROW_CLASS_NAME = {
-    inventory: 'product-manage-item',
-};
-
-export const TABLE_CELL_CLASS_NAME = {
-    inventory: {
-        productName: 'product-manage-name',
-        productPrice: 'product-manage-price',
-        productQuantity: 'product-manage-quantity',
-    },
-};
-
-export const TABLE_TITLE = {
-    productName: '상품명',
-    productPrice: '가격',
-    productQuantity: '수량',
-};
